modify structure of directories in app
