+++ src/components/Navigation.js
@@ -1,7 +1,8 @@
 import { useContext } from 'react';
-import { UserContext } from '../controllers/UserContext';
 import { Link } from 'react-router-dom';
+import { UserContext } from '../controllers/UserContext';
 import { auth } from '../index';
+import { signOut } from '@firebase/auth';
 import { styled, alpha } from '@mui/material/styles';
 import AppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
@@ -18,7 +19,6 @@ import LogoutOutlinedIcon from '@mui/icons-material/LogoutOutlined';
 import PersonAddIcon from '@mui/icons-material/PersonAdd';
 import AccountCircleOutlinedIcon from '@mui/icons-material/AccountCircleOutlined';
 import logo from '../logo.png';
-import { signOut } from '@firebase/auth';
 
 const Navigation = () => {
     const Search = styled('div')(({ theme }) => ({
