+++ src/components/auxiliaries/Spinner.js
@@ -1,4 +1,4 @@
-import '../index.css'
+import '../../index.css';
 
 const Spinner = () => {
     return (
