+++ src/components/books/BookDetails.js
@@ -2,8 +2,8 @@ import { useState, useEffect } from 'react';
 import { useParams, useHistory } from 'react-router-dom';
 import { doc, getDoc } from 'firebase/firestore';
 import { ref, getDownloadURL } from 'firebase/storage';
-import { db, storage } from '../index';
-import Spinner from './Spinner';
+import { db, storage } from '../../index';
+import Spinner from '../auxiliaries/Spinner';
 import Box from '@mui/material/Box';
 import CardMedia from '@mui/material/CardMedia';
 import AddShoppingCartIcon from '@mui/icons-material/AddShoppingCart';
