+++ src/components/forms/SignUpForm.js
@@ -1,28 +1,31 @@
 import { useState } from 'react';
 import { Redirect, Link } from 'react-router-dom';
-import { signInWithEmailAndPassword } from 'firebase/auth';
-import { auth } from './../index';
+import { auth, db } from '../../index';
+import { createUserWithEmailAndPassword } from "firebase/auth";
+import { doc, setDoc } from "firebase/firestore";
+import { useStyles } from './Form.styled';
 import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
 import Container from '@mui/material/Container';
 import SendIcon from '@mui/icons-material/Send';
 import TextField from '@mui/material/TextField';
-import { useStyles } from './Form.styled';
 
-const SignInForm = () => {
+const SignUpForm = () => {
 
     const classes = useStyles();
 
     const [formData, setFormData] = useState({
         email: '',
-        password: ''
+        password: '',
+        username: ''
     })
-    const { email, password } = formData;
+    const { email, password, username } = formData;
 
     const [shouldRedirect, setShouldRedirect] = useState(false); 
 
     const [emailError, setEmailError] = useState(false);
     const [passwordError, setPasswordError] = useState(false);
+    const [usernameError, setUsernameError] = useState(false);
 
     const handleChange = e => {
         setFormData({
@@ -35,19 +38,25 @@ const SignInForm = () => {
         e.preventDefault();
         setEmailError(false);
         setPasswordError(false);
+        setUsernameError(false);
         if (email === '') {
             setEmailError(true);
         } 
         if (password === '') {
             setPasswordError(true);
         }
-        if (email && password) {
-            signInWithEmailAndPassword(auth, email, password)
+        if (username === '') {
+            setUsernameError(true);
+        }
+        if (email && password && username) {
+            createUserWithEmailAndPassword(auth, email, password)
                 .then(userCredential => {
-                    setFormData({
-                    email: '',
-                    password: ''
-                    });
+                    const userRef = doc(db, 'users', userCredential.user.uid)
+                    setDoc(userRef, {
+                        username
+                    })
+                })
+                .then(() => {
                     setShouldRedirect(true);
                 })
                 .catch(error => {
@@ -66,27 +75,27 @@ const SignInForm = () => {
         <Container>
             <div className={classes.wrapper}>
                 <form 
-                    autoComplete="off"
-                    className={classes.form}
+                    autoComplete="off" 
+                    className={classes.form} 
                     onSubmit={handleSubmit}
                 >
                     <div>
-                        <Typography 
-                            variant="h5"
-                            component="h2"
-                            gutterBottom
-                            align="center"
-                        >
-                            Welcome!
-                        </Typography>
-                        <Typography 
-                            variant="subtitle1"
-                            component="h2"
-                            gutterBottom
-                            align="center"
-                        >
-                            Sign in and enjoy!
-                        </Typography>
+                    <Typography 
+                        variant="h5"
+                        component="h2"
+                        gutterBottom
+                        align="center"
+                    >
+                        Welcome!
+                    </Typography>
+                    <Typography 
+                        variant="subtitle1"
+                        component="h2"
+                        gutterBottom
+                        align="center"
+                    >
+                        Do not wait register now and join our community!
+                    </Typography>
                     </div>
                     <TextField 
                         label="email address"
@@ -114,6 +123,19 @@ const SignInForm = () => {
                         error={passwordError}
                         onChange={handleChange}
                     />
+                    <TextField 
+                        label="username"
+                        variant="outlined"
+                        color="primary"
+                        type="text"
+                        name="username"
+                        value={username}
+                        fullWidth
+                        required
+                        className={classes.field}
+                        error={usernameError}
+                        onChange={handleChange}
+                    />
                     <Button
                         variant="contained"
                         color="primary"
@@ -121,16 +143,16 @@ const SignInForm = () => {
                         endIcon={<SendIcon />} 
                         className={classes.btn} 
                     >
-                        SIGN IN
+                        SIGN UP
                     </Button>
                 </form>
                     <Typography
                         variant="overline"
                         align="right"    
                     >
-                        New to Smart Books? 
-                        <Link to="/sign-up" className={classes.link}>
-                            Create new account
+                        Already have an account?
+                        <Link to="/sign-in" className={classes.link}>
+                            Sign in
                         </Link>
                     </Typography>
             </div>
@@ -138,4 +160,4 @@ const SignInForm = () => {
     );
 }
 
-export default SignInForm;
+export default SignUpForm;
