+++ src/firebase.config.js
@@ -7,3 +7,4 @@ export const firebaseConfig = {
   appId: "1:249816505082:web:a9ac5920bcb97a80e10743",
   measurementId: "G-TPFR8QK4FM"
 };
+
